Move integration tests into the "it" folder
