+++ langchain4j-watsonx/src/test/java/dev/langchain4j/model/watsonx/it/WatsonxChatModelListenerIT.java
@@ -1,4 +1,4 @@
-package dev.langchain4j.model.watsonx;
+package dev.langchain4j.model.watsonx.it;
 
 import com.ibm.watsonx.ai.chat.ChatService;
 import com.ibm.watsonx.ai.core.auth.AuthenticationProvider;
@@ -8,6 +8,7 @@ import dev.langchain4j.model.chat.ChatModel;
 import dev.langchain4j.model.chat.common.AbstractChatModelListenerIT;
 import dev.langchain4j.model.chat.listener.ChatModelListener;
 import dev.langchain4j.model.chat.request.ChatRequestParameters;
+import dev.langchain4j.model.watsonx.WatsonxChatModel;
 import java.time.Duration;
 import java.util.List;
 import org.junit.jupiter.api.condition.EnabledIfEnvironmentVariable;
